profile로직 Redux로 분리
