+++ src/hooks/useUserProject.js
@@ -1,18 +1,18 @@
 import { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-import { saveProject } from "../store/projectSlice";
 import {
   getProjectListAction,
   getSelectedProjectAction,
 } from "../store/thunkAction/projectAction";
-import { getUserProjectList, getProjectDetails } from "../utils/API";
 import { ProfileTypes } from "../constants";
 
 export default function useUserProject(projectId) {
   const [dsn, setDsn] = useState(projectId);
-  const [profiles, setProfiles] = useState({});
-  const { projectList, selectedProject } = useSelector(state => state.project);
+  // const [profiles, setProfiles] = useState({});
+  const { projectList, selectedProject, profile } = useSelector(
+    state => state.project,
+  );
   const dispatch = useDispatch();
 
   useEffect(() => {
@@ -33,33 +33,33 @@ export default function useUserProject(projectId) {
     dispatch(getSelectedProjectAction(dsn));
   };
 
-  const getProjectProfiles = () => {
-    const profile = {
-      "first-input": [],
-      "largest-contentful-paint": [],
-      "layout-shift": [],
-      longtask: [],
-      navigation: [],
-      paint: [],
-    };
+  // const getProjectProfiles = () => {
+  //   const profile = {
+  //     "first-input": [],
+  //     "largest-contentful-paint": [],
+  //     "layout-shift": [],
+  //     longtask: [],
+  //     navigation: [],
+  //     paint: [],
+  //   };
 
-    if (!selectedProject?.performance) return;
-    const profileList = selectedProject?.performance;
+  //   if (!selectedProject?.performance) return;
+  //   const profileList = selectedProject?.performance;
 
-    for (let i = 0; i < profileList.length; i++) {
-      const key = Object.keys(profileList[i])[0];
-      if (ProfileTypes.includes(key)) {
-        profile[key].push(profileList[i][key]);
-      }
-    }
+  //   for (let i = 0; i < profileList.length; i++) {
+  //     const key = Object.keys(profileList[i])[0];
+  //     if (ProfileTypes.includes(key)) {
+  //       profile[key].push(profileList[i][key]);
+  //     }
+  //   }
 
-    setProfiles(profile);
-  };
+  //   setProfiles(profile);
+  // };
 
   return {
     getSelectedProject,
     getUserProject,
-    setProfiles,
+    // setProfiles,
     projectList,
     dsn,
     setDsn,
